+++ code/modules/antagonists/bloodsucker/items/bloodsucker_organs.dm
@@ -5,7 +5,8 @@
 	var/obj/item/organ/O
 	// Heart
 	O = owner.current.getorganslot(ORGAN_SLOT_HEART)
-	if(!istype(O, /obj/item/organ/heart/vampheart))
+
+	if(!istype(O, /obj/item/organ/heart/vampheart) && !istype(O, /obj/item/organ/heart/demon))
 		qdel(O)
 		var/obj/item/organ/heart/vampheart/H = new
 		H.Insert(owner.current)
@@ -29,10 +30,6 @@
 /obj/item/organ/heart/vampheart
 	beating = 0
 	var/fakingit = 0
-
-/obj/item/organ/heart/vampheart/prepare_eat()
-	..()
-	// Do cool stuff for eating vamp heart?
 
 /obj/item/organ/heart/vampheart/Restart()
 	beating = 0	// DONT run ..(). We don't want to start beating again.
